Standardize tenant API platform-state test fixtures
The tenant API test surfaces had started importing fake clients and fixture wiring directly out of test_tenant_api_handler.py, which made unrelated test modules depend on one monolithic test file. This change extracts the shared fake platform-state seam into a dedicated support module and rewires the ops and module-level suites to use the same builders and helpers.

Constraint: Shared test support had to preserve current tenant API behavior across handler, ops, and module dispatch tests
Rejected: Keep importing fake helpers from test_tenant_api_handler.py | Preserved cross-test coupling and made fixture drift harder to control
Rejected: Move every test fake into conftest.py | Broadened fixture visibility beyond the tenant API test slice without need
Confidence: high
Scope-risk: narrow
Reversibility: clean
Directive: Add shared tenant API test doubles to tests/unit/tenant_api_test_support.py first; do not reintroduce cross-imports from one test file into another
Tested: uv run pytest tests/unit/test_tenant_api_handler.py tests/unit/test_ops_api.py tests/unit/test_tenant_api_modules.py -q; uv run ruff check tests/unit/tenant_api_test_support.py tests/unit/test_tenant_api_handler.py tests/unit/test_ops_api.py tests/unit/test_tenant_api_modules.py; make validate-local; make preflight-session; make pre-validate-session
Not-tested: Full unit suite outside the tenant API-related test surfaces
